+++ src/main/java/indiv/abko/taskflow/domain/auth/service/WithdrawUseCase.java
@@ -7,8 +7,8 @@ import indiv.abko.taskflow.domain.auth.dto.command.WithdrawCommand;
 import indiv.abko.taskflow.domain.auth.exception.AuthErrorCode;
 import indiv.abko.taskflow.domain.user.entity.Member;
 import indiv.abko.taskflow.domain.user.service.MemberServiceApi;
+import indiv.abko.taskflow.global.auth.JwtBlacklist;
 import indiv.abko.taskflow.global.exception.BusinessException;
-import indiv.abko.taskflow.global.jwt.JwtBlacklistService;
 import lombok.RequiredArgsConstructor;
 
 @Service
@@ -17,7 +17,7 @@ public class WithdrawUseCase {
 
 	private final MemberServiceApi memberService;
 	private final PasswordEncoder passwordEncoder;
-	private final JwtBlacklistService jwtBlacklistService;
+	private final JwtBlacklist jwtBlacklist;
 
 	public void execute(WithdrawCommand command) {
 		Member member = memberService.getByIdOrThrow(command.memberId());
@@ -26,7 +26,7 @@ public class WithdrawUseCase {
 			throw new BusinessException(AuthErrorCode.PASSWORD_MISMATCH);
 		}
 
-		jwtBlacklistService.addToBlacklist(command.token());
+		jwtBlacklist.addToBlacklist(command.token());
 
 		memberService.withdraw(member);
 	}
